dashboard/requirements.txt
streamlit
pandas
altair
boto3
python-dotenv